Introduce explicit classes for task types
