+++ optiframe/__init__.py
@@ -1,6 +1,6 @@
 from optiframe.workflow_engine import Task, StepData
 from optiframe.framework import Optimizer
 from optiframe.framework.errors import InfeasibleError
-from optiframe.framework.tasks import SolutionObjValue
+from optiframe.framework.default_tasks import SolutionObjValue
 
 __all__ = ["Task", "StepData", "Optimizer", "InfeasibleError", "SolutionObjValue"]
